:recycle: refactor(pipeline): Rename 'Pipeline' to 'Pipelines' namespace
